#6 more or less final emotion detection
This classifier is trained on several data sets that contain text and a true or false value  that mark whether or not the text contains a emotion. These sets are news headlines (all false), Quotes from the tv show friends pulled from another study (these were all true), and some of our own labeled data (mixed true and false). This had a cross val score of ~91%. 

I applied this to some of my own hand labeled data (111 true tweets, 12 negative). At first this was receiving a high precision, but very low accuracy. This is due to how unbalanced this data is. However, when re sampling the training data and retraining the classifier on a ratio more similar to that of my test data, the accuracy rose to 90%